Made hero movement a bit snappier
Hero rotates more quickly and only moves in exactly the direction you
are pushing on the keyboard (allows for more precise movement, at the
cost of a small amount of realism).  Hero does not currently regenerate
health, and a check has been added in Hero for when health reaches 0.
Created a NextLevel function to eventually be used for cycling through
levels.